restructing FetchBook
